// src/models/types/location.ts
enum Location {
  BESAID = "BESAID",
  BIKANEL = "BIKANEL",
  CALM_LANDS = "CALM_LANDS",
  DJOSE_ROAD = "DJOSE_ROAD",
  INSIDE_SIN = "INSIDE_SIN",
  KILIKA = "KILIKA",
  MACALANIA = "MACALANIA",
  MI_IHEN_HIGHROAD = "MI_IHEN_HIGHROAD",
  MONT_GAGAZET = "MONT_GAGAZET",
  MUSHROOM_ROCK_ROAD = "MUSHROOM_ROCK_ROAD",
  OMEGA_DUNGEON = "OMEGA_DUNGEON",
  STOLEN_FAYTH_CAVERN = "STOLEN_FAYTH_CAVERN",
  THUNDER_PLAINS = "THUNDER_PLAINS",
}

export { Location };